animation bords moved to animationController
